feat(dev): decouple UI primitives for gallery
